NoteSchool v0.0.4
*Arreglo de bugs en el horario.
*Arreglo de interfaces graficas para horario.

+++ NoteSchool/Layout/Horario/ScHorario.cs
@@ -98,22 +98,44 @@ namespace NoteSchool.Layout.Horario {
             ScEditarClase scEditarClase = new ScEditarClase();
 
             //Obtener el id del dia.
-            String id = dataGVHorario.CurrentRow.Cells["idhorario_" + lDia.Text].Value.ToString();
-            
-            //Metodo para settear el parametro dia y el id.
-            scEditarClase.setDia(lDia.Text);
-            scEditarClase.setId(id);
+            if (dataGVHorario.CurrentRow != null) {
 
-            scEditarClase.Show(); //Mostrar el form para editar.
+               String id = dataGVHorario.CurrentRow.Cells["idhorario_" + lDia.Text].Value.ToString();
+                
+               //Metodo para settear el parametro dia y el id.
+               scEditarClase.setDia(lDia.Text);
+               scEditarClase.setId(id);
+
+               scEditarClase.Show(); //Mostrar el form para editar.
+
+            } else {
+
+                MessageBox.Show("Selecciona una clase.");
+            
+            }
+            
         }
 
         //Evento click para eliminar la clase.
         private void butEliminarClase_Click(object sender, EventArgs e) {
 
-            String id = dataGVHorario.CurrentRow.Cells["idhorario_" + lDia.Text].Value.ToString();
-            hr.eliminarClase(id, lDia.Text);
+            if (dataGVHorario.CurrentRow != null) {
+
+               String id = dataGVHorario.CurrentRow.Cells["idhorario_" + lDia.Text].Value.ToString();
+               hr.eliminarClase(id, lDia.Text);
+
+            } else {
+
+                MessageBox.Show("Selecciona una clase.");
+            
+            }
 
         }
-    
+
+        //Evento click para actualizar la base de datos.
+        private void butActualizar_Click(object sender, EventArgs e) {
+            hr.cargarHorario(dataGVHorario, "horario_" + lDia.Text);
+        }
+
     }
 }
